PLAT-2707: create Google Artifacty Registry (#73)
* PLAT-2707: Simplify GKE terraform

* PLAT-2707: Google Artifact Repositories

create an artifact repository for models and environments
create a service account to read/write the repositories

grant forge access to PUSH images

* PLAT-2707: Allow Pull from all Nodes

Set correct oauth scope. This fixes the previous GPU issue as well
as granting the default behavior of pulling from Container or Artifact
Registry.

* PLAT-2707: containerd is the default

Now that we are using Artifact Registry, use containerd as the default
container runtime.